Modification of the SDK to accepte some future fonctionnalties to add…

+++ UMI3D-SDK/Assets/UMI3D SDK/EnvironmentDevelopmentKit/UserCapture/Runtime/Animations/HandPoseSetter.cs
@@ -217,45 +217,45 @@ namespace umi3d.edk.userCapture
 
                 HandPose.PhalanxRotations.Clear();
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightThumbProximal, nameof(BoneType.RightThumbProximal), ScriptableHand.Get(nameof(BoneType.RightThumbProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightThumbIntermediate, nameof(BoneType.RightThumbIntermediate), ScriptableHand.Get(nameof(BoneType.RightThumbIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightThumbDistal, nameof(BoneType.RightThumbDistal), ScriptableHand.Get(nameof(BoneType.RightThumbDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightThumbProximal, nameof(BoneType.RightThumbProximal), ScriptableHand.Get(nameof(BoneType.RightThumbProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightThumbIntermediate, nameof(BoneType.RightThumbIntermediate), ScriptableHand.Get(nameof(BoneType.RightThumbIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightThumbDistal, nameof(BoneType.RightThumbDistal), ScriptableHand.Get(nameof(BoneType.RightThumbDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightIndexProximal, nameof(BoneType.RightIndexProximal), ScriptableHand.Get(nameof(BoneType.RightIndexProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightIndexIntermediate, nameof(BoneType.RightIndexIntermediate), ScriptableHand.Get(nameof(BoneType.RightIndexIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightIndexDistal, nameof(BoneType.RightIndexDistal), ScriptableHand.Get(nameof(BoneType.RightIndexDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightIndexProximal, nameof(BoneType.RightIndexProximal), ScriptableHand.Get(nameof(BoneType.RightIndexProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightIndexIntermediate, nameof(BoneType.RightIndexIntermediate), ScriptableHand.Get(nameof(BoneType.RightIndexIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightIndexDistal, nameof(BoneType.RightIndexDistal), ScriptableHand.Get(nameof(BoneType.RightIndexDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightMiddleProximal, nameof(BoneType.RightMiddleProximal), ScriptableHand.Get(nameof(BoneType.RightMiddleProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightMiddleIntermediate, nameof(BoneType.RightMiddleIntermediate), ScriptableHand.Get(nameof(BoneType.RightMiddleIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightMiddleDistal, nameof(BoneType.RightMiddleDistal), ScriptableHand.Get(nameof(BoneType.RightMiddleDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightMiddleProximal, nameof(BoneType.RightMiddleProximal), ScriptableHand.Get(nameof(BoneType.RightMiddleProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightMiddleIntermediate, nameof(BoneType.RightMiddleIntermediate), ScriptableHand.Get(nameof(BoneType.RightMiddleIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightMiddleDistal, nameof(BoneType.RightMiddleDistal), ScriptableHand.Get(nameof(BoneType.RightMiddleDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightRingProximal, nameof(BoneType.RightRingProximal), ScriptableHand.Get(nameof(BoneType.RightRingProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightRingIntermediate, nameof(BoneType.RightRingIntermediate), ScriptableHand.Get(nameof(BoneType.RightRingIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightRingDistal, nameof(BoneType.RightRingDistal), ScriptableHand.Get(nameof(BoneType.RightRingDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightRingProximal, nameof(BoneType.RightRingProximal), ScriptableHand.Get(nameof(BoneType.RightRingProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightRingIntermediate, nameof(BoneType.RightRingIntermediate), ScriptableHand.Get(nameof(BoneType.RightRingIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightRingDistal, nameof(BoneType.RightRingDistal), ScriptableHand.Get(nameof(BoneType.RightRingDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightLittleProximal, nameof(BoneType.RightLittleProximal), ScriptableHand.Get(nameof(BoneType.RightLittleProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightLittleIntermediate, nameof(BoneType.RightLittleIntermediate), ScriptableHand.Get(nameof(BoneType.RightLittleIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.RightLittleDistal, nameof(BoneType.RightLittleDistal), ScriptableHand.Get(nameof(BoneType.RightLittleDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightLittleProximal, nameof(BoneType.RightLittleProximal), ScriptableHand.Get(nameof(BoneType.RightLittleProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightLittleIntermediate, nameof(BoneType.RightLittleIntermediate), ScriptableHand.Get(nameof(BoneType.RightLittleIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.RightLittleDistal, nameof(BoneType.RightLittleDistal), ScriptableHand.Get(nameof(BoneType.RightLittleDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftThumbProximal, nameof(BoneType.LeftThumbProximal), ScriptableHand.Get(nameof(BoneType.LeftThumbProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftThumbIntermediate, nameof(BoneType.LeftThumbIntermediate), ScriptableHand.Get(nameof(BoneType.LeftThumbIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftThumbDistal, nameof(BoneType.LeftThumbDistal), ScriptableHand.Get(nameof(BoneType.LeftThumbDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftThumbProximal, nameof(BoneType.LeftThumbProximal), ScriptableHand.Get(nameof(BoneType.LeftThumbProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftThumbIntermediate, nameof(BoneType.LeftThumbIntermediate), ScriptableHand.Get(nameof(BoneType.LeftThumbIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftThumbDistal, nameof(BoneType.LeftThumbDistal), ScriptableHand.Get(nameof(BoneType.LeftThumbDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftIndexProximal, nameof(BoneType.LeftIndexProximal), ScriptableHand.Get(nameof(BoneType.LeftIndexProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftIndexIntermediate, nameof(BoneType.LeftIndexIntermediate), ScriptableHand.Get(nameof(BoneType.LeftIndexIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftIndexDistal, nameof(BoneType.LeftIndexDistal), ScriptableHand.Get(nameof(BoneType.LeftIndexDistal)).Rot)); ;
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftIndexProximal, nameof(BoneType.LeftIndexProximal), ScriptableHand.Get(nameof(BoneType.LeftIndexProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftIndexIntermediate, nameof(BoneType.LeftIndexIntermediate), ScriptableHand.Get(nameof(BoneType.LeftIndexIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftIndexDistal, nameof(BoneType.LeftIndexDistal), ScriptableHand.Get(nameof(BoneType.LeftIndexDistal)).Rot)); ;
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftMiddleProximal, nameof(BoneType.LeftMiddleProximal), ScriptableHand.Get(nameof(BoneType.LeftMiddleProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftMiddleIntermediate, nameof(BoneType.LeftMiddleIntermediate), ScriptableHand.Get(nameof(BoneType.LeftMiddleIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftMiddleDistal, nameof(BoneType.LeftMiddleDistal), ScriptableHand.Get(nameof(BoneType.LeftMiddleDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftMiddleProximal, nameof(BoneType.LeftMiddleProximal), ScriptableHand.Get(nameof(BoneType.LeftMiddleProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftMiddleIntermediate, nameof(BoneType.LeftMiddleIntermediate), ScriptableHand.Get(nameof(BoneType.LeftMiddleIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftMiddleDistal, nameof(BoneType.LeftMiddleDistal), ScriptableHand.Get(nameof(BoneType.LeftMiddleDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftRingProximal, nameof(BoneType.LeftRingProximal), ScriptableHand.Get(nameof(BoneType.LeftRingProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftRingIntermediate, nameof(BoneType.LeftRingIntermediate), ScriptableHand.Get(nameof(BoneType.LeftRingIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftRingDistal, nameof(BoneType.LeftRingDistal), ScriptableHand.Get(nameof(BoneType.LeftRingDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftRingProximal, nameof(BoneType.LeftRingProximal), ScriptableHand.Get(nameof(BoneType.LeftRingProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftRingIntermediate, nameof(BoneType.LeftRingIntermediate), ScriptableHand.Get(nameof(BoneType.LeftRingIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftRingDistal, nameof(BoneType.LeftRingDistal), ScriptableHand.Get(nameof(BoneType.LeftRingDistal)).Rot));
 
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftLittleProximal, nameof(BoneType.LeftLittleProximal), ScriptableHand.Get(nameof(BoneType.LeftLittleProximal)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftLittleIntermediate, nameof(BoneType.LeftLittleIntermediate), ScriptableHand.Get(nameof(BoneType.LeftLittleIntermediate)).Rot));
-                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation(BoneType.LeftLittleDistal, nameof(BoneType.LeftLittleDistal), ScriptableHand.Get(nameof(BoneType.LeftLittleDistal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftLittleProximal, nameof(BoneType.LeftLittleProximal), ScriptableHand.Get(nameof(BoneType.LeftLittleProximal)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftLittleIntermediate, nameof(BoneType.LeftLittleIntermediate), ScriptableHand.Get(nameof(BoneType.LeftLittleIntermediate)).Rot));
+                HandPose.PhalanxRotations.Add(new UMI3DHandPose.PhalanxRotation((int)BoneType.LeftLittleDistal, nameof(BoneType.LeftLittleDistal), ScriptableHand.Get(nameof(BoneType.LeftLittleDistal)).Rot));
             }
 
             EditorUtility.SetDirty(HandPose);
